Change argument type to uint16

+++ ledcircle.h
@@ -1,6 +1,6 @@
 #ifndef LEDCIRCLE_H
 #define LEDCIRCLE_H
 
-void ledcircle_select(uint8_t led);
+void ledcircle_select(uint16_t led);
 
 #endif
